use u64 instead of i32

+++ src/bench/mod.rs
@@ -51,7 +51,7 @@ fn format_duration(duration: Duration) -> String {
     }
 }
 
-fn generate_random_array(size: usize, min: i32, max: i32, seed: Option<u64>) -> Vec<i32> {
+fn generate_random_array(size: usize, min: u64, max: u64, seed: Option<u64>) -> Vec<u64> {
     let mut rng = match seed {
         Some(seed_value) => StdRng::seed_from_u64(seed_value),
         None => StdRng::from_entropy(),
@@ -60,12 +60,12 @@ fn generate_random_array(size: usize, min: i32, max: i32, seed: Option<u64>) -> 
     (0..size).map(|_| rng.gen_range(min..=max)).collect()
 }
 
-pub fn bench_algorithm_with_random_u64<F>(array_size: usize, sort_fn: F, num: i32, algoname: &str)
+pub fn bench_algorithm_with_random_u64<F>(array_size: usize, sort_fn: F, num: u64, algoname: &str)
 where
-    F: Fn(&mut Vec<i32>),
+    F: Fn(&mut Vec<u64>),
 {
     // Ensure algo works :|
-    let mut vec_to_sort = generate_random_array(array_size, 0, array_size as i32, None);
+    let mut vec_to_sort = generate_random_array(array_size, 0, array_size as u64, None);
     sort_fn(&mut vec_to_sort);
     assert!(
         vec_to_sort.is_sorted(),
@@ -76,7 +76,7 @@ where
     // Bench time
     let mut acc = Duration::ZERO;
     for _ in 0..num {
-        let mut vec_to_sort = generate_random_array(array_size, 0, array_size as i32, None);
+        let mut vec_to_sort = generate_random_array(array_size, 0, array_size as u64, None);
         let start = Instant::now();
         sort_fn(&mut vec_to_sort);
         acc += start.elapsed();
